[337152] [Mongo] Create a MongoDBStore https://bugs.eclipse.org/bugs/show_bug.cgi?id=337152

+++ plugins/org.eclipse.emf.cdo.tests.mongodb/src/org/eclipse/emf/cdo/tests/mongodb/MongoDBStoreRepositoryConfig.java
@@ -62,7 +62,10 @@ public class MongoDBStoreRepositoryConfig extends RepositoryConfig
   public IStore createStore(String repoName)
   {
     MongoURI mongoURI = new MongoURI("mongodb://localhost");
-    dropDatabase(mongoURI, repoName);
+    if (!isRestarting())
+    {
+      dropDatabase(mongoURI, repoName);
+    }
 
     return CDOMongoDBUtil.createStore(mongoURI, repoName);
   }
